+++ ckb-mercury-sdk/src/main/java/org/nervos/mercury/model/req/ToInfo.java
@@ -1,10 +1,12 @@
 package org.nervos.mercury.model.req;
 
+import java.math.BigInteger;
+
 public class ToInfo {
   public String address;
-  public long amount;
+  public BigInteger amount;
 
-  public ToInfo(String address, long amount) {
+  public ToInfo(String address, BigInteger amount) {
     this.address = address;
     this.amount = amount;
   }
