+++ ckb-mercury-sdk/src/main/java/org/nervos/mercury/model/resp/BalanceResponse.java
@@ -2,10 +2,12 @@ package org.nervos.mercury.model.resp;
 
 import org.nervos.mercury.model.common.AssetInfo;
 
+import java.math.BigInteger;
+
 public class BalanceResponse {
   public String ownership;
   public AssetInfo assetInfo;
-  public long occupied;
-  public long free;
-  public long frozen;
+  public BigInteger occupied;
+  public BigInteger free;
+  public BigInteger frozen;
 }
