+++ ckb-mercury-sdk/src/test/java/mercury/BuildSimpleTransferTransactionTest.java
@@ -8,6 +8,7 @@ import org.nervos.mercury.model.SimpleTransferPayloadBuilder;
 import org.nervos.mercury.model.common.AssetInfo;
 
 import java.io.IOException;
+import java.math.BigInteger;
 
 public class BuildSimpleTransferTransactionTest {
 
@@ -17,7 +18,7 @@ public class BuildSimpleTransferTransactionTest {
     builder.assetInfo(AssetInfo.newCkbAsset());
     builder.addFrom("ckt1qzda0cr08m85hc8jlnfp3zer7xulejywt49kt2rr0vthywaa50xwsqfqyerlanzmnkxtmd9ww9n7gr66k8jt4tclm9jnk");
     builder.addTo("ckt1qzda0cr08m85hc8jlnfp3zer7xulejywt49kt2rr0vthywaa50xwsqg958atl2zdh8jn3ch8lc72nt0cf864ecqdxm9zf"
-        , 10000000000L);
+        , BigInteger.valueOf(10000000000L));
     builder.feeRate(500L);
 
     TransactionWithScriptGroups s =
