+++ ckb-mercury-sdk/src/test/java/mercury/BuildTransferTransactionTest.java
@@ -10,6 +10,7 @@ import org.nervos.mercury.model.req.item.ItemFactory;
 import org.nervos.mercury.model.req.payload.CapacityProvider;
 
 import java.io.IOException;
+import java.math.BigInteger;
 
 public class BuildTransferTransactionTest {
 
@@ -19,7 +20,7 @@ public class BuildTransferTransactionTest {
     builder.assetInfo(AssetInfo.newCkbAsset());
     builder.addFrom(ItemFactory.newAddressItem("ckt1qzda0cr08m85hc8jlnfp3zer7xulejywt49kt2rr0vthywaa50xwsqfqyerlanzmnkxtmd9ww9n7gr66k8jt4tclm9jnk"));
     builder.addTo("ckt1qzda0cr08m85hc8jlnfp3zer7xulejywt49kt2rr0vthywaa50xwsqg958atl2zdh8jn3ch8lc72nt0cf864ecqdxm9zf"
-        , 100);
+        , BigInteger.valueOf(100));
     builder.payFee(CapacityProvider.FROM);
     builder.feeRate(1100L);
